CARDS-1768: Refactoring/UX improvement: replace long 'Select's with Autocomplete ExportButton: move the questionnaire autocomplete functionality into a separate reusable component
